// Tp10/LabAngular/src/app/models/Employee.ts
export class Employee{
    FirstName: string;
    LastName: string;
    City: string;
    Country: string;
    EmployeeID: number;
    Title: string;
    TitleOfCourtesy: string;
    BirthDate: string;
    HireDate: string;
    Address: string;
    Region:string;
    PostalCode: string;
    HomePhone: string;
    Extension: string;
    Notes: string;
    ReportsTo: number;
    PhotoPath: string;

}